+++ packages/gx-agent-cli/CHANGELOG.md
@@ -3,6 +3,19 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0](https://github.com/Sphereon/gx-agent/compare/v0.1.2...v0.2.0) (2023-01-27)
+
+### Bug Fixes
+
+- fixed issue in vp issue ([acfd0e6](https://github.com/Sphereon/gx-agent/commit/acfd0e6b88c51574ba1f216de3c408e36f24db78))
+- readme updated ([ce399ff](https://github.com/Sphereon/gx-agent/commit/ce399ffbb3fdb5f4a84fa3e353f252026950b927))
+
+### Features
+
+- add ecosystem config support ([b90b410](https://github.com/Sphereon/gx-agent/commit/b90b410b6dc5d410e5bbe44def03cd937e72aba6))
+- add ecosystem documentation ([00d5292](https://github.com/Sphereon/gx-agent/commit/00d52921457d637befc69f54bcd50aa96e78f7a7))
+- added onboarding functionality for ecosystem ([7dd0fa9](https://github.com/Sphereon/gx-agent/commit/7dd0fa9ab3ce969684d3b351bd9d7ebf96e766df))
+
 ## [0.1.2](https://github.com/Sphereon/gx-agent/compare/v0.1.1...v0.1.2) (2023-01-26)
 
 **Note:** Version bump only for package @sphereon/gx-agent-cli